requirements.txt
flask
torch
torchvision
torchaudio
diffusers
Pillow
requests
